Implemented get_code, get_byte and at_end_of_stream. Recognized strange bug which occurs at disjunctions in the interpreter.
I think there is a bad assertion in builtin/control.py, I removed it (line 72 in fail()).
This solved the problem which occured for example while executing the following kind of code:

	1 = 1, (2 = 3; 3 = 4).

This code fails now as expected, but the following code breaks the interpreter:
(blub is an empty file.)

	open('blub', read, S), (at_end_of_sream(S); at_end_of_stream(S)).

Didn't figure out, what exactly is the problem here, because

	open('blub', read, S), (at_end_of_sream(S); 1 = 2).

doesn't cause any problems.