Fix the standard order.
Note that we do not use the iso standard order, but rather an order more akin
to that of SWI prolog. As it stands the order is:

Var < Float = Number = BigInt < Atom < Term

Then numbers are tested by value, however, mixed int/float comparisons
are *always* compared as floats.

Note that strings are not yet supported and thus do not appear in the order.

Removed some broken tests in favour of some more systematic and
comprehensive tests.

OK cfbolz.